+++ crates/hotshot/Cargo.toml
@@ -54,7 +54,6 @@ time = { workspace = true }
 tracing = { workspace = true }
 vbs = { workspace = true }
 jf-signature.workspace = true
-hotshot-orchestrator = { path = "../orchestrator" }
 blake3.workspace = true
 sha2 = { workspace = true }
 url = { workspace = true }
