+++ crates/types/src/lib.rs
@@ -23,8 +23,10 @@ pub mod constants;
 pub mod data;
 pub mod error;
 pub mod event;
+pub mod hotshot_config_file;
 pub mod light_client;
 pub mod message;
+pub mod network;
 pub mod qc;
 pub mod request_response;
 pub mod signature_key;
@@ -32,7 +34,9 @@ pub mod simple_certificate;
 pub mod simple_vote;
 pub mod stake_table;
 pub mod traits;
+pub mod upgrade_config;
 pub mod utils;
+pub mod validator_config;
 pub mod vid;
 pub mod vote;
