+++ crates/types/src/network.rs
@@ -13,6 +13,7 @@ use std::{
     vec,
 };
 
+use crate::hotshot_config_file::HotShotConfigFile;
 use clap::ValueEnum;
 use hotshot_types::{
     constants::REQUEST_DATA_DELAY, light_client::StateVerKey, traits::signature_key::SignatureKey,
@@ -26,22 +27,20 @@ use toml;
 use tracing::{error, info};
 use vec1::Vec1;
 
-use crate::client::OrchestratorClient;
+/// default number of rounds to run
+pub const ORCHESTRATOR_DEFAULT_NUM_ROUNDS: usize = 100;
+/// default number of transactions per round
+pub const ORCHESTRATOR_DEFAULT_TRANSACTIONS_PER_ROUND: usize = 10;
+/// default size of transactions
+pub const ORCHESTRATOR_DEFAULT_TRANSACTION_SIZE: usize = 100;
+/// default delay before beginning consensus
+pub const ORCHESTRATOR_DEFAULT_START_DELAY_SECONDS: u64 = 60;
 
 /// Configuration describing a libp2p node
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
 pub struct Libp2pConfig {
     /// The bootstrap nodes to connect to (multiaddress, serialized public key)
     pub bootstrap_nodes: Vec<(PeerId, Multiaddr)>,
-}
-
-/// configuration for a web server
-#[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
-pub struct WebServerConfig {
-    /// the url to run on
-    pub url: Url,
-    /// the time to wait between polls
-    pub wait_between_polls: Duration,
 }
 
 /// configuration for combined network
@@ -181,79 +180,10 @@ pub enum NetworkConfigSource {
 }
 
 impl<K: SignatureKey> NetworkConfig<K> {
-    /// Asynchronously retrieves a `NetworkConfig` either from a file or from an orchestrator.
-    ///
-    /// This function takes an `OrchestratorClient`, an optional file path, and Libp2p-specific parameters.
-    ///
-    /// If a file path is provided, the function will first attempt to load the `NetworkConfig` from the file.
-    /// If the file does not exist or cannot be read, the function will fall back to retrieving the `NetworkConfig` from the orchestrator.
-    /// In this case, if the path to the file does not exist, it will be created.
-    /// The retrieved `NetworkConfig` is then saved back to the file for future use.
-    ///
-    /// If no file path is provided, the function will directly retrieve the `NetworkConfig` from the orchestrator.
-    ///
-    /// # Errors
-    /// If we were unable to load the configuration.
-    ///
-    /// # Arguments
-    ///
-    /// * `client` - An `OrchestratorClient` used to retrieve the `NetworkConfig` from the orchestrator.
-    /// * `identity` - A string representing the identity for which to retrieve the `NetworkConfig`.
-    /// * `file` - An optional string representing the path to the file from which to load the `NetworkConfig`.
-    /// * `libp2p_address` - An optional address specifying where other Libp2p nodes can reach us
-    /// * `libp2p_public_key` - The public key in which other Libp2p nodes can reach us with
-    ///
-    /// # Returns
-    ///
-    /// This function returns a tuple containing a `NetworkConfig` and a `NetworkConfigSource`. The `NetworkConfigSource` indicates whether the `NetworkConfig` was loaded from a file or retrieved from the orchestrator.
-    pub async fn from_file_or_orchestrator(
-        client: &OrchestratorClient,
-        file: Option<String>,
-        libp2p_advertise_address: Option<Multiaddr>,
-        libp2p_public_key: Option<PeerId>,
-    ) -> anyhow::Result<(NetworkConfig<K>, NetworkConfigSource)> {
-        if let Some(file) = file {
-            info!("Retrieving config from the file");
-            // if we pass in file, try there first
-            match Self::from_file(file.clone()) {
-                Ok(config) => Ok((config, NetworkConfigSource::File)),
-                Err(e) => {
-                    // fallback to orchestrator
-                    error!("{e}, falling back to orchestrator");
-
-                    let config = client
-                        .get_config_without_peer(libp2p_advertise_address, libp2p_public_key)
-                        .await?;
-
-                    // save to file if we fell back
-                    if let Err(e) = config.to_file(file) {
-                        error!("{e}");
-                    };
-
-                    Ok((config, NetworkConfigSource::File))
-                }
-            }
-        } else {
-            info!("Retrieving config from the orchestrator");
-
-            // otherwise just get from orchestrator
-            Ok((
-                client
-                    .get_config_without_peer(libp2p_advertise_address, libp2p_public_key)
-                    .await?,
-                NetworkConfigSource::Orchestrator,
-            ))
-        }
-    }
-
     /// Get a temporary node index for generating a validator config
-    pub async fn generate_init_validator_config(
-        client: &OrchestratorClient,
-        is_da: bool,
-    ) -> ValidatorConfig<K> {
+    pub fn generate_init_validator_config(cur_node_index: u16, is_da: bool) -> ValidatorConfig<K> {
         // This cur_node_index is only used for key pair generation, it's not bound with the node,
-        // lather the node with the generated key pair will get a new node_index from orchestrator.
-        let cur_node_index = client.get_node_index_for_init_validator_config().await;
+        // later the node with the generated key pair will get a new node_index from orchestrator.
         ValidatorConfig::generated_from_seed_indexed([0u8; 32], cur_node_index.into(), 1, is_da)
     }
 
@@ -500,255 +430,6 @@ impl<K: SignatureKey> From<NetworkConfigFile<K>> for NetworkConfig<K> {
             builder: val.builder,
             random_builder: val.random_builder,
             public_keys: val.public_keys,
-        }
-    }
-}
-
-/// Default builder URL, used as placeholder
-fn default_builder_urls() -> Vec1<Url> {
-    vec1::vec1![Url::parse("http://0.0.0.0:3311").unwrap()]
-}
-
-/// Holds configuration for a `HotShot`
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(bound(deserialize = ""))]
-pub struct HotShotConfigFile<KEY: SignatureKey> {
-    /// The proportion of nodes required before the orchestrator issues the ready signal,
-    /// expressed as (numerator, denominator)
-    pub start_threshold: (u64, u64),
-    /// Total number of staked nodes in the network
-    pub num_nodes_with_stake: NonZeroUsize,
-    #[serde(skip)]
-    /// My own public key, secret key, stake value
-    pub my_own_validator_config: ValidatorConfig<KEY>,
-    #[serde(skip)]
-    /// The known nodes' public key and stake value
-    pub known_nodes_with_stake: Vec<PeerConfig<KEY>>,
-    #[serde(skip)]
-    /// The known DA nodes' public key and stake values
-    pub known_da_nodes: Vec<PeerConfig<KEY>>,
-    #[serde(skip)]
-    /// The known non-staking nodes'
-    pub known_nodes_without_stake: Vec<KEY>,
-    /// Number of staking DA nodes
-    pub staked_da_nodes: usize,
-    /// Number of fixed leaders for GPU VID
-    pub fixed_leader_for_gpuvid: usize,
-    /// Base duration for next-view timeout, in milliseconds
-    pub next_view_timeout: u64,
-    /// Duration for view sync round timeout
-    pub view_sync_timeout: Duration,
-    /// The exponential backoff ration for the next-view timeout
-    pub timeout_ratio: (u64, u64),
-    /// The delay a leader inserts before starting pre-commit, in milliseconds
-    pub round_start_delay: u64,
-    /// Delay after init before starting consensus, in milliseconds
-    pub start_delay: u64,
-    /// Number of network bootstrap nodes
-    pub num_bootstrap: usize,
-    /// The maximum amount of time a leader can wait to get a block from a builder
-    pub builder_timeout: Duration,
-    /// Time to wait until we request data associated with a proposal
-    pub data_request_delay: Option<Duration>,
-    /// Builder API base URL
-    #[serde(default = "default_builder_urls")]
-    pub builder_urls: Vec1<Url>,
-    /// Upgrade config
-    pub upgrade: UpgradeConfig,
-}
-
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-#[serde(bound(deserialize = ""))]
-/// Holds configuration for the upgrade task.
-pub struct UpgradeConfig {
-    /// View to start proposing an upgrade
-    pub start_proposing_view: u64,
-    /// View to stop proposing an upgrade. To prevent proposing an upgrade, set stop_proposing_view <= start_proposing_view.
-    pub stop_proposing_view: u64,
-    /// View to start voting on an upgrade
-    pub start_voting_view: u64,
-    /// View to stop voting on an upgrade. To prevent voting on an upgrade, set stop_voting_view <= start_voting_view.
-    pub stop_voting_view: u64,
-    /// Unix time in seconds at which we start proposing an upgrade
-    pub start_proposing_time: u64,
-    /// Unix time in seconds at which we stop proposing an upgrade. To prevent proposing an upgrade, set stop_proposing_time <= start_proposing_time.
-    pub stop_proposing_time: u64,
-    /// Unix time in seconds at which we start voting on an upgrade
-    pub start_voting_time: u64,
-    /// Unix time in seconds at which we stop voting on an upgrade. To prevent voting on an upgrade, set stop_voting_time <= start_voting_time.
-    pub stop_voting_time: u64,
-}
-
-// Explicitly implementing `Default` for clarity.
-#[allow(clippy::derivable_impls)]
-impl Default for UpgradeConfig {
-    fn default() -> Self {
-        UpgradeConfig {
-            start_proposing_view: u64::MAX,
-            stop_proposing_view: 0,
-            start_voting_view: u64::MAX,
-            stop_voting_view: 0,
-            start_proposing_time: u64::MAX,
-            stop_proposing_time: 0,
-            start_voting_time: u64::MAX,
-            stop_voting_time: 0,
-        }
-    }
-}
-
-/// Holds configuration for a validator node
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, Default)]
-#[serde(bound(deserialize = ""))]
-pub struct ValidatorConfigFile {
-    /// The validator's seed
-    pub seed: [u8; 32],
-    /// The validator's index, which can be treated as another input to the seed
-    pub node_id: u64,
-    // The validator's stake, commented for now
-    // pub stake_value: u64,
-    /// Whether or not we are DA
-    pub is_da: bool,
-}
-
-impl ValidatorConfigFile {
-    /// read the validator config from a file
-    /// # Panics
-    /// Panics if unable to get the current working directory
-    pub fn from_file(dir_str: &str) -> Self {
-        let current_working_dir = match env::current_dir() {
-            Ok(dir) => dir,
-            Err(e) => {
-                error!("get_current_working_dir error: {:?}", e);
-                PathBuf::from("")
-            }
-        };
-        let filename =
-            current_working_dir.into_os_string().into_string().unwrap() + "/../../" + dir_str;
-        match fs::read_to_string(filename.clone()) {
-            // If successful return the files text as `contents`.
-            Ok(contents) => {
-                let data: ValidatorConfigFile = match toml::from_str(&contents) {
-                    // If successful, return data as `Data` struct.
-                    // `d` is a local variable.
-                    Ok(d) => d,
-                    // Handle the `error` case.
-                    Err(e) => {
-                        // Write `msg` to `stderr`.
-                        error!("Unable to load data from `{}`: {}", filename, e);
-                        ValidatorConfigFile::default()
-                    }
-                };
-                data
-            }
-            // Handle the `error` case.
-            Err(e) => {
-                // Write `msg` to `stderr`.
-                error!("Could not read file `{}`: {}", filename, e);
-                ValidatorConfigFile::default()
-            }
-        }
-    }
-}
-
-impl<KEY: SignatureKey> From<HotShotConfigFile<KEY>> for HotShotConfig<KEY> {
-    fn from(val: HotShotConfigFile<KEY>) -> Self {
-        HotShotConfig {
-            execution_type: ExecutionType::Continuous,
-            start_threshold: val.start_threshold,
-            num_nodes_with_stake: val.num_nodes_with_stake,
-            known_da_nodes: val.known_da_nodes,
-            known_nodes_with_stake: val.known_nodes_with_stake,
-            known_nodes_without_stake: val.known_nodes_without_stake,
-            my_own_validator_config: val.my_own_validator_config,
-            da_staked_committee_size: val.staked_da_nodes,
-            fixed_leader_for_gpuvid: val.fixed_leader_for_gpuvid,
-            next_view_timeout: val.next_view_timeout,
-            view_sync_timeout: val.view_sync_timeout,
-            timeout_ratio: val.timeout_ratio,
-            round_start_delay: val.round_start_delay,
-            start_delay: val.start_delay,
-            num_bootstrap: val.num_bootstrap,
-            builder_timeout: val.builder_timeout,
-            data_request_delay: val
-                .data_request_delay
-                .unwrap_or(Duration::from_millis(REQUEST_DATA_DELAY)),
-            builder_urls: val.builder_urls,
-            start_proposing_view: val.upgrade.start_proposing_view,
-            stop_proposing_view: val.upgrade.stop_proposing_view,
-            start_voting_view: val.upgrade.start_voting_view,
-            stop_voting_view: val.upgrade.stop_voting_view,
-            start_proposing_time: val.upgrade.start_proposing_time,
-            stop_proposing_time: val.upgrade.stop_proposing_time,
-            start_voting_time: val.upgrade.start_voting_time,
-            stop_voting_time: val.upgrade.stop_voting_time,
-        }
-    }
-}
-/// default number of rounds to run
-pub const ORCHESTRATOR_DEFAULT_NUM_ROUNDS: usize = 100;
-/// default number of transactions per round
-pub const ORCHESTRATOR_DEFAULT_TRANSACTIONS_PER_ROUND: usize = 10;
-/// default size of transactions
-pub const ORCHESTRATOR_DEFAULT_TRANSACTION_SIZE: usize = 100;
-/// default delay before beginning consensus
-pub const ORCHESTRATOR_DEFAULT_START_DELAY_SECONDS: u64 = 60;
-
-impl<K: SignatureKey> From<ValidatorConfigFile> for ValidatorConfig<K> {
-    fn from(val: ValidatorConfigFile) -> Self {
-        // here stake_value is set to 1, since we don't input stake_value from ValidatorConfigFile for now
-        ValidatorConfig::generated_from_seed_indexed(val.seed, val.node_id, 1, val.is_da)
-    }
-}
-impl<KEY: SignatureKey> From<ValidatorConfigFile> for HotShotConfig<KEY> {
-    fn from(value: ValidatorConfigFile) -> Self {
-        let mut config: HotShotConfig<KEY> = HotShotConfigFile::default().into();
-        config.my_own_validator_config = value.into();
-        config
-    }
-}
-
-impl<KEY: SignatureKey> Default for HotShotConfigFile<KEY> {
-    fn default() -> Self {
-        // The default number of nodes is 5
-        let staked_da_nodes: usize = 5;
-
-        // Aggregate the DA nodes
-        let mut known_da_nodes = Vec::new();
-
-        let gen_known_nodes_with_stake = (0..10)
-            .map(|node_id| {
-                let mut cur_validator_config: ValidatorConfig<KEY> =
-                    ValidatorConfig::generated_from_seed_indexed([0u8; 32], node_id, 1, false);
-
-                // Add to DA nodes based on index
-                if node_id < staked_da_nodes as u64 {
-                    known_da_nodes.push(cur_validator_config.public_config());
-                    cur_validator_config.is_da = true;
-                }
-
-                cur_validator_config.public_config()
-            })
-            .collect();
-
-        Self {
-            num_nodes_with_stake: NonZeroUsize::new(10).unwrap(),
-            start_threshold: (1, 1),
-            my_own_validator_config: ValidatorConfig::default(),
-            known_nodes_with_stake: gen_known_nodes_with_stake,
-            known_nodes_without_stake: vec![],
-            staked_da_nodes,
-            known_da_nodes,
-            fixed_leader_for_gpuvid: 1,
-            next_view_timeout: 10000,
-            view_sync_timeout: Duration::from_millis(1000),
-            timeout_ratio: (11, 10),
-            round_start_delay: 1,
-            start_delay: 1,
-            num_bootstrap: 5,
-            builder_timeout: Duration::from_secs(10),
-            data_request_delay: Some(Duration::from_millis(REQUEST_DATA_DELAY)),
-            builder_urls: default_builder_urls(),
-            upgrade: UpgradeConfig::default(),
         }
     }
 }
